src/requirements.txt
amqp==2.4.0
astroid==2.1.0
billiard==3.5.0.5
bokeh==1.0.4
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-celery-results==1.0.4
django-redis==4.10.0
djangorestframework==3.9.1
idna==2.8
isort==4.3.4
Jinja2==2.10
kombu==4.2.2.post1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
mongoengine==0.16.3
numpy==1.16.0
packaging==19.0
pandas==0.24.0
Pillow==5.4.1
pylint==2.2.2
pymongo==3.7.2
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
PyYAML>=4.2b1
redis==3.1.0
requests==2.21.0
selenium==3.141.0
six==1.12.0
tornado==5.1.1
typed-ast==1.3.1
urllib3==1.24.1
vine==1.2.0
wrapt==1.11.1
flower==0.9.2